+++ src/effects/mod.rs
@@ -4,8 +4,8 @@ mod greyscale;
 
 pub use blur::Blur;
 pub use dither::{
-    Color, Colour, DiffusionAlgorithm, DiffusionKernel, DiffusionScan, DiffusionTap,
-    ErrorDiffusion, NoiseAlgorithm, NoiseDither, OrderedDither, Palette, Threshold, ThresholdMap,
-    ThresholdRotation,
+    Color, Colour, DiffusionAlgorithm, DiffusionKernel, DiffusionScan, DiffusionTap, DotShape,
+    ErrorDiffusion, Halftone, HalftoneShape, NoiseAlgorithm, NoiseDither, OrderedDither, Palette,
+    Threshold, ThresholdMap, ThresholdRotation,
 };
 pub use greyscale::Greyscale;